Introduce syntax for splicing whole values in
This is inspired by Pulumi YAML's syntax: when an expression is _just_
a reference, like the value in

    foo: ${var}

then substitute the value straight in, rather than trying to construct
a string. If the ref is only part of the expression, serialise the
value and use that as part of the output string.

(Usually you'd expect people to line things up so they are
substituting atoms into strings, and other kinds of value into whole
fields.)

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>