Provide machinery for compiling templates
The template given at the end of a comprehension can be "pretreated",
since it doesn't depend on any of the variables. The scheme I've
implemented here is to look through the template, and find all the
places that a value is replaced with the result of an expression,
collecting funcs that will do the replacement as it goes.

This has the benefit of being simple; it's pretty brittle otherwise
though, because it mutates (and returns!) the original value. This is
correct as far as it goes, because the values in question will always
be replaced, so you don't see old values.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>