Return a copy of the interpolated blank
If template.evaluate is used to build a list of results, then mutating
the "blank" value is a problem -- all the list items are the same
value.

As a simple way to avoid that, deep copy the blank after doing
interpolation, and return the copy. As before, it doesn't matter that
the blank is mutated, as the next interpolation will replace all the
interpolated-in values.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>