Rearrange API type and rewrite eval to suit
This commit makes changes to the Comprehension API type, to reflect
the non-nested grammar, and rewrites `evalTop` to deal with the new
structure.

The rewrite uses `compileTemplate` rather than the interpolation
funcs, so those latter can now be removed.

`type env struct` loses its `value` field, because it's never used as
a dynamic environment; in fact, all I really need is a list of
variable names, but I'm keeping `env` around for now because I think
it might be useful when there are expressions in more places.

Instatiating all the values is fun. Since the generators are
independent, I can run them all once each. Then it just takes a bit of
recursion to construct each vector of values and run the template.

The examples follow the structure of the API of course, so they get
translated.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>