Flatten template outputs if a list, in controller
It's handy to be able to give a set of objects to create (e.g., a
Deployment and a Service). The most natural way is to just supply a
list of those objects in the template. So: when the controller sees a
list as the output of the template, it'll create each item in the
list.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>