Simplify recursion
Instead of treating one generator left as a special case and iterating
through its values, just recurse again. Then the special case is also
the base case -- no more generators -- and all that remains is to
instatiate the template, which is simpler. (I also corrected a mistake
in the non-special case: it should skip a value when the filter is
_not_ truthy).

This comes at a cost, though, which is that since the recursive part
wants to concatenate slices (`append(out, more...)`), the base case
must return a slice of one item.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>