Avoid append of single-item slices
The classic treatment of list comprehensions given in "The
Implementation of Functional Programming Languages" (Simon
Peyton-Jones, 1987) uses a program transformation to avoid using
append, which is expensive with cons-based lists. It does this by
passing a base list down the recursions, and consing on the way back
up.

Though appending to slices is less expensive, I can use the same trick
to avoid the awkwardness of constructing single-item slices just so I
can append them.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>