Prepare for eval'ing generator expressions
This prepares for generators which include expressions themselves,
e.g., the second generator expression below, which uses the value from
the first to construct its list:

```
for:
- var: a
  in:
    list: [1,2,3]
- var: b
  in:
    list:
    - ${a + 1}
    - ${a + 2}
```

The main change here is to delay evaluation of the values until
bindings are ready to evaluate them. Efficiency for fixed lists can be
recovered by noting that they do not have any embedded expressions,
and can be calculated once.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>