Move eval machinery to internal/eval
This is so I can use it from the command-line tool, as well as the
controller.

I didn't want to use the ginkgo setup in the new package, so I wrote a
TestMain which starts up the envtest Environment, instead, and rewrote
the tests in generate_test.go as `*testing.T` tests.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>