Primitive command-line tool
There's nothing especially stopping a command-line tool from running
comprehensions (either in the filesystem, or in a cluster); and,
printing out the results (or creating things in a cluster).

This does the bare bones, with no frills: read in a file as a
comprehension object, run it and print the results.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>