Make an HTTP request and use the result as a list
This commit does two related things:

- adds a `request` generator, which (crudely, for now) does an HTTP
  request to the URL given and parses the result as JSON;

- changes the `list` generator type from `[]apiextensions.JSON` to
  `*apiextensions.JSON`, so that it can be a literal list, a list with
  interpolations, _or_ an expression into which a list can be
  interpolated.

The motivation for these changes is to be able to have comprehensions
like:

```
for:
- var: result
  in:
    request:
      url: https://api.github.com/repos/org/repo/pulls
- var: pr
  in:
    list: ${result}
```

The response to the GitHub API request is a JSON list; the second
qualifier in the expression lifts that response into a list, to be
iterated over. This wasn't possible before because the field was typed
as a list.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>