Implement interpolation into request generator
Since I need `replaceStrPointer` for replacing the string-valued
fields in a `Request`, I moved that and its cousin out.

There's two levels of test: first, that using the generator with
static values; second, that using generator with values interpolated
in works. Strictly speaking the second covers the first, but I don't
know, strength in numbers.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>