Avoid failing if an object already exists
If I use k8sClient.Create only, it'll fail the second time it runs.
Instead, use controllerutils.CreateOrUpdate, which will update in the
case that the object already exists. The test added checks that
rerunning a comprehension (by changing its spec) succeeds.

Using CreateOrUpdate with unstructured.Unstructured is subtle and I'm
not sure I've got it right. With a typed object you can pick the bit
to update -- .Spec, usually -- and just set that, to make sure you're
not stepping on values set by the Kubernetes API server, or a
webhook. I have made an approximation, which is to set the fields as
generated.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>